+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -60,6 +60,7 @@ import {
   IndividualDataSchema,
   isErrorWithMessage,
   isObject,
+  isType,
   ProcessStatus,
   setCollectionFlowStatus,
   TWorkflowHelpers,
@@ -87,6 +88,7 @@ import {
 import {
   ApprovalState,
   BusinessPosition,
+  CreatedFrom,
   Customer,
   EndUser,
   Prisma,
@@ -122,6 +124,8 @@ import { PartialDeep } from 'type-fest';
 import { WorkflowAssignee, WorkflowRuntimeListItemModel } from './workflow-runtime-list-item.model';
 import { formatIndividualVerification } from '@/common/utils/idv';
 import { AssessmentsService } from '@/assessments/assessments.service';
+import { KycService } from '@/kyc/kyc.service';
+import z from 'zod';
 
 type TEntityId = string;
 
@@ -139,8 +143,9 @@ const getAvatarUrl = (website: string | undefined | null) =>
     ? `https://t2.gstatic.com/faviconV2?client=SOCIAL&type=FAVICON&fallback_opts=TYPE,SIZE,URL&url=${website}&size=40`
     : null;
 
-const handlePromiseAll = async <T>(promises: Record<string, Promise<T>>) => {
-  const record: Record<string, T> = {};
+const handlePromiseAll = async <TPromises extends Record<string, Promise<any>>>(
+  promises: TPromises,
+) => {
   const errors: Array<{ key: string; error: unknown }> = [];
   const promisesEntries = Object.entries(promises);
   const results = await Promise.all(
@@ -155,7 +160,8 @@ const handlePromiseAll = async <T>(promises: Record<string, Promise<T>>) => {
     ),
   );
   const fulfilledResults = results.filter(
-    (res): res is { key: string; status: 'fulfilled'; data: T } => res.status === 'fulfilled',
+    (res): res is { key: string; status: 'fulfilled'; data: Awaited<TPromises[keyof TPromises]> } =>
+      res.status === 'fulfilled',
   );
   const rejectedResults = results.filter(
     (res): res is { key: string; status: 'rejected'; error: unknown } => res.status === 'rejected',
@@ -174,11 +180,16 @@ const handlePromiseAll = async <T>(promises: Record<string, Promise<T>>) => {
     );
   }
 
-  for (const fulfilledResult of fulfilledResults) {
-    record[fulfilledResult.key] = fulfilledResult.data;
-  }
+  return fulfilledResults.reduce(
+    (acc, fulfilledResult) => {
+      acc[fulfilledResult.key as keyof TPromises] = fulfilledResult.data;
 
-  return record;
+      return acc;
+    },
+    {} as {
+      [TKey in keyof TPromises]: Awaited<TPromises[TKey]>;
+    },
+  );
 };
 
 @Injectable()
@@ -209,6 +220,7 @@ export class WorkflowService {
     private readonly storageService: StorageService,
     private readonly workflowLogService: WorkflowLogService,
     private readonly assessmentsService: AssessmentsService,
+    private readonly kycService: KycService,
   ) {}
 
   async createWorkflowDefinition(data: WorkflowDefinitionCreateDto) {
@@ -2364,6 +2376,88 @@ export class WorkflowService {
             },
           },
         );
+      });
+
+      service.subscribe('RUN_AML_ON_REGISTRY_PEOPLE_OF_INTEREST', async ({ payload }) => {
+        const PayloadWithPeopleOfInterestSchema = z.object({
+          peopleOfInterest: z.array(
+            z.object({
+              firstName: z.string(),
+              lastName: z.string(),
+              role: z.string(),
+            }),
+          ),
+        });
+        const checkIsValidPayload = isType(PayloadWithPeopleOfInterestSchema);
+
+        if (!checkIsValidPayload(payload)) {
+          this.logger.log('Skipping AML on registry people of interest');
+
+          return;
+        }
+
+        const callbackUrl = `${env.APP_API_URL}/api/v1/external/workflows/${workflowRuntimeData.id}/hook/NO_OP?processName=aml-unified-api`;
+        let peopleOfInterest: Array<{
+          ballerineEntityId: string;
+          firstName: string;
+          lastName: string;
+          role: string;
+        }> = [];
+
+        const promises: Record<string, Promise<any>> = {};
+
+        for (const personOfInterest of payload.peopleOfInterest) {
+          const { customer, endUser } = await handlePromiseAll({
+            customer: this.customerService.getByProjectId(currentProjectId),
+            endUser: this.endUserService.create(
+              {
+                data: {
+                  firstName: personOfInterest.firstName,
+                  lastName: personOfInterest.lastName,
+                  createdFrom: CreatedFrom.registry,
+                  projectId: currentProjectId,
+                },
+              },
+              transaction,
+            ),
+          });
+
+          peopleOfInterest.push({
+            ballerineEntityId: endUser.id,
+            firstName: endUser.firstName,
+            lastName: endUser.lastName,
+            role: personOfInterest.role,
+          });
+
+          promises[endUser.id] = this.kycService.initiateAml({
+            endUserId: endUser.id,
+            clientId: customer.name,
+            vendor: 'veriff',
+            immediateResults: true,
+            ongoingMonitoring: false,
+            callbackUrl,
+            firstName: endUser.firstName,
+            lastName: endUser.lastName,
+          });
+        }
+
+        await handlePromiseAll(promises);
+
+        await service.sendEvent({
+          type: BUILT_IN_EVENT.DEEP_MERGE_CONTEXT,
+          payload: {
+            arrayMergeOption: ARRAY_MERGE_OPTION.BY_INDEX,
+            newContext: {
+              entity: {
+                data: {
+                  additionalInfo: {
+                    peopleOfInterest,
+                  },
+                },
+              },
+            },
+          },
+        });
       });
 
       if (!service.getSnapshot().nextEvents.includes(type)) {